+++ src/components/NftPreview/NftPreview.tsx
